requirements.txt
dependency_injector
dataclasses-json
marshmallow
pymongo
bs4
lxml
paho-mqtt
pymongo
